Use size_t to keep the value returned by strlen().
When using a variable to store the result of strlen(), make sure its
type is size_t just like the function returns, rather than using a
smaller type like int.

+++ src/input/libreal/real.c
@@ -313,7 +313,7 @@ static void calc_response_string (char *result, char *challenge) {
 
 void real_calc_response_and_checksum (char *response, char *chksum, char *challenge) {
 
-  int   ch_len, resp_len;
+  size_t ch_len, resp_len;
   int   i;
   char *ptr;
   char  buf[128];
